Pushed stochastic opacity into the BVH
For scenes with lot's of opacity masks this is a performance win and solves the issue with the infinite loop nicely. Requires some extra code, though.
With this cange we formalized that opacity masks have to be images (not procedural textures). This simplifies things a little, but also seems somwhat arbitrary.