Introduced ShapeSampler concept that is needed by lights now
There is some clunkiness, including the name, but it already solves a long standing problem when several lights share an image mapped material but have different shapes.

It also solves the mesh-light "variant" approach in a more general way.

Hopefully this could serve to somehow store portal light resources, if we get to it.